chore: update package version to 1.3.0 and dependencies
feat: export additional modules from index.ts

refactor: move Razorpay types to a separate types.ts file

refactor: simplify razorpay.ts by removing redundant code and exporting Razorpay class

feat: implement useRazorpay hook in use-razorpay.ts for script loading
