+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "razorpay-hook",
-  "version": "1.2.3",
+  "version": "1.3.0",
   "description": "Seamlessly integrate Razorpay's powerful payment solutions into your React and Next.js applications. This package simplifies the process of setting up secure, reliable, and efficient payment gateways, giving developers a streamlined experience while building high-performance apps.",
   "main": "dist/index.js",
   "module": "dist/index.mjs",
@@ -32,8 +32,8 @@
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^12.1.1",
     "@types/react": "^18.3.12",
-    "next": "^15.2.4",
-    "react": "^19.0.0",
+    "next": "16.0.5",
+    "react": "19.2.0",
     "rollup": "^4.26.0",
     "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
