+++ src/index.ts
@@ -1,2 +1,3 @@
-
-export * from './razorpay'
+export * from "./razorpay";
+export * from "./types";
+export * from "./use-razorpay";
